Bugfixes and cleanups
* Add 'BIOS' to the end of the release title if tagged BIOS in No-Intro
* Sizes are now represented as i64
* Parse as Result<Vec<Result<GameEntry>>> to allow partial failures of parsing

+++ src/shiratsu-lib/src/parse/common/rom_entry.rs
@@ -9,7 +9,7 @@ pub struct RomEntry {
     /// The canonical file name of the ROM
     pub(in super::super) file_name: String,
     /// The size of the ROM
-    pub(in super::super) size: u32,
+    pub(in super::super) size: i64,
 }
 
 /// Describes a single file that is a part of a GameEntry
@@ -31,7 +31,7 @@ impl RomEntry {
         &self.file_name
     }
     /// The size of the ROM
-    pub fn size(&self) -> u32 {
+    pub fn size(&self) -> i64 {
         self.size
     }
 }
